Finish separating code into VolumePage and TimeControl
VolumePage has all code regarding retrieving the IDs of timepoint folders
TimeControl has all the code regarding loading timepoints which includes:
 - Vuetify elements like the slider, play/pause, next, etc.
 - Retrieving the IDs for the data files for a specific timepoint
